+++ packages/sdk/src/transactionManager.ts
@@ -4,9 +4,11 @@ import { PrepareParams, CancelParams, FulfillParams, getUuid, isNode } from "@co
 import { TransactionManager as TTransactionManager, IERC20Minimal } from "@connext/nxtp-contracts/typechain";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
 import ERC20 from "@connext/nxtp-contracts/artifacts/contracts/interfaces/IERC20Minimal.sol/IERC20Minimal.json";
-import contractDeployments from "@connext/nxtp-contracts/deployments.json";
+import _contractDeployments from "@connext/nxtp-contracts/deployments.json";
 
 import { ChainNotConfigured } from "./error";
+
+export const contractDeployments = _contractDeployments;
 
 /**
  * Returns the address of the `TransactionManager` deployed to the provided chain, or undefined if it has not been deployed
